BOOST_PARAMETER_NAME refactoring
Moved shareable support facilities to separate header in anticipation of bringing BOOST_PARAMETER_NESTED_KEYWORD home from Boost.Accumulators in the future.
